split the iterator tests into a new file
